Changed the OOP model to have an LED and cube Class
This might be flawed as I need a way to set which LED is controlled with which pin as each led is controlled with two pins I think that I probally need to re-evaluate what I am doing. LEDs are not the lowest level of abstraction I have a whole load of circutary to consider.